fix: refactor queue tests
