+++ internal/testsuite/submitter/submitter.go
@@ -18,14 +18,8 @@ import (
 	"github.com/armadaproject/armada/pkg/client"
 )
 
-// submitRetryInterval is the delay between submit retries when the queue is not yet
-// visible to the server (the queue cache refreshes periodically, so a just-created
-// queue can briefly be unrecognised by SubmitJobs).
+// Retry to account for recently created queues returning not found errors on SubmitJobs
 const submitRetryInterval = 500 * time.Millisecond
-
-// submitRetryTimeout bounds how long we keep retrying a submission that fails because
-// the queue can't be found. It should comfortably exceed the server's queue cache
-// refresh period.
 const submitRetryTimeout = 15 * time.Second
 
 var submissionSerializer sync.Mutex
@@ -174,12 +168,10 @@ retryLoop:
 		case <-time.After(submitRetryInterval):
 		}
 	}
-	srv.logSubmitStatus(req, err, time.Now().Sub(start))
+	srv.logSubmitStatus(req, err, time.Since(start))
 	return res, err
 }
 
-// isQueueNotFoundErr returns true if err is the transient PermissionDenied returned by
-// SubmitJobs when a recently created queue hasn't yet reached the server's queue cache.
 func isQueueNotFoundErr(err error) bool {
 	s, ok := status.FromError(err)
 	if !ok {
