+++ testsuite/testcases/queue/README.md
@@ -2,44 +2,39 @@
 
 Test cases in this folder exercise queue lifecycle behavior (create, update,
 assert, delete) via `queueConfig` on the `TestSpec` (see
-`pkg/api/testspec.proto`). `queueConfig` has four sections, run in this order:
+`pkg/api/testspec.proto`). `queueConfig` has three sections, run in this order:
 
 1. `setup` — create queue(s) before the rest of the test runs.
 2. `update` — apply an update to the created queue(s).
 3. `assertions` — check queue state.
-4. `teardown` — delete the queue(s) (runs last, always, even on failure).
+
+Queues created by `setup` (or `TestSpec.queue`, if `setup` wasn't used) are
+always deleted at the end of the test, even on failure.
 
 ## setup
 
 ```yaml
 queueConfig:
-  setup:
-    create: true       # create the queue(s) named by TestSpec.queue
-    randomSuffix: true  # append a random suffix to avoid name collisions
-    numBatches: 5        # number of batches of queues to create (default 1)
-    batchSize: 5         # queues created per batch (default 1)
-    interval: "1s"        # time between batches (default: as fast as possible)
-    queueSpecs:            # optional templates (see below)
-      - priorityFactor: 2.0
+    setup: # presence of `setup` creates the queue(s) named by TestSpec.queue
+        numBatches: 5 # number of batches of queues to create (default 1)
+        batchSize: 5 # queues created per batch (default 1)
+        interval: "1s" # time between batches (default: as fast as possible)
+        queueSpec: # optional template (see below)
+            priorityFactor: 2.0
 ```
 
-If `numBatches == 1` and `batchSize == 1` (the default), exactly one queue is
-created, named `TestSpec.queue` (or `<queue>-<randomSuffix>` if
-`randomSuffix` is set). Otherwise queues are named `<queue>-<index>`.
-
-`queueSpecs` lets a test create multiple queues with different properties
-(anything on `api.Queue` — priority factor, permissions, labels, etc.) in one
-batch. One copy of each spec is created per queue slot in the batch. When
-more than one queue is created in a batch, the batched
-`POST /v1/batched/create_queues` endpoint is used instead of individual
-`CreateQueue` calls.
+`queueSpec` lets a test create queues with non-default properties (anything
+on `api.Queue` — priority factor, permissions, labels, etc.). One copy of the
+spec is created per queue slot in the batch. When more than one queue is
+created in a batch, the batched `POST /v1/batched/create_queues` endpoint is
+used instead of individual `CreateQueue` calls.
 
 ## update
 
 ```yaml
 queueConfig:
-  update:
-    priorityFactor: 2.0
+    update:
+        priorityFactor: 2.0
 ```
 
 `update` is an `api.Queue` template applied to every queue created by
@@ -55,34 +50,22 @@ A list of checks, run in order after `update`:
 
 ```yaml
 queueConfig:
-  assertions:
-    - activeInPool: "default"       # queue appears in GetActiveQueues for this pool
-    - notActiveInPool: "default"    # queue does NOT appear in GetActiveQueues for this pool
-    - appearsInStream: true          # queue appears in the GetQueues streaming response
-    - matches:                        # GetQueue's result matches this template
-        priorityFactor: 2.0           # (useful for verifying an `update` was applied)
-    - deleted: true                    # after teardown deletes the queue(s), GetQueue returns NOT_FOUND
+    assertions:
+        - activeInPool: "default" # queue appears in GetActiveQueues for this pool
+        - notActiveInPool: "default" # queue does NOT appear in GetActiveQueues for this pool
+        - appearsInStream: true # queue appears in the GetQueues streaming response
+        - matches: # GetQueue's result matches this template
+              priorityFactor: 2.0 # (useful for verifying an `update` was applied)
+        - deleted: true # after the queue(s) are deleted at the end of the test, GetQueue returns NOT_FOUND
 ```
 
-`deleted` only takes effect once `teardown` actually deletes the queue(s) —
-it's evaluated as part of teardown, not before it.
-
-## teardown
-
-```yaml
-queueConfig:
-  teardown:
-    skipDelete: true      # don't delete the queue(s) after the test (default: delete)
-    expectNotFound: true   # expect DeleteQueue to return NOT_FOUND (tests the error path)
-```
-
-Queues are deleted by default at the end of every test that uses
-`queueConfig`; set `skipDelete: true` to opt out.
+`deleted` is evaluated after the queue(s) are deleted at the end of the test,
+not before.
 
 ## Examples
 
 - `crud_1x1.yaml` — create, update, assert, delete a single queue.
-- `crud_batch_5x5.yaml` — create 25 queues (5 batches of 5) via `queueSpecs`,
+- `crud_batch_5x5.yaml` — create 25 queues (5 batches of 5) via `queueSpec`,
   update and assert them, exercising the batched endpoints.
 - `active_list_1x1.yaml` — create a queue, submit jobs to it, assert it
   appears in `GetActiveQueues`.
